Module stm32f103xx::dma1::ccr1

DMA channel configuration register (DMA_CCR)

Structs

CIRCR
Value of the field

DIRR
Value of the field

HTIER
Value of the field

MEM2MEMR
Value of the field

MINCR
Value of the field

PINCR
Value of the field

R
Value read from the register

TCIER
Value of the field

TEIER
Value of the field

W
Value to write to the register

_CIRCW
Proxy

_DIRW
Proxy

_ENW
Proxy

_HTIEW
Proxy

_MEM2MEMW
Proxy

_MINCW
Proxy

_MSIZEW
Proxy

_PINCW
Proxy

_PLW
Proxy

_PSIZEW
Proxy

_TCIEW
Proxy

_TEIEW
Proxy

Enums

PLR
Possible values of the field PL

PLW
Values that can be written to the field PL

PSIZER
Possible values of the field PSIZE

PSIZEW
Values that can be written to the field PSIZE

Type Definitions

ENR
Possible values of the field EN

ENW
Values that can be written to the field EN

MSIZER
Possible values of the field MSIZE

MSIZEW
Values that can be written to the field MSIZE