Enum stm32f103xx::dma1::ccr1::PLR

pub enum PLR {
    LOW,
    MEDIUM,
    HIGH,
    VERYHIGH,
}

Possible values of the field PL

Variants
LOW

Low

MEDIUM

Medium

HIGH

High

VERYHIGH

Very High

Methods

impl PLR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_low(&self) -> bool

Checks if the value of the field is LOW

pub fn is_medium(&self) -> bool

Checks if the value of the field is MEDIUM

pub fn is_high(&self) -> bool

Checks if the value of the field is HIGH

pub fn is_very_high(&self) -> bool

Checks if the value of the field is VERYHIGH