Enum stm32f103xx::flash::acr::LATENCYR

pub enum LATENCYR {
    ZERO,
    ONE,
    TWO,
    _Reserved(u8),
}

Possible values of the field LATENCY

Variants
ZERO

Zero wait state, if 0hz SYSCLK to 24 MHz

ONE

One wait state, if 24 MHz SYSCLK to 48 MHz

TWO

Two wait states, if 48 MHz SYSCLK to 72 MHz

_Reserved(u8)

Reserved

Methods

impl LATENCYR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_zero(&self) -> bool

Checks if the value of the field is ZERO

pub fn is_one(&self) -> bool

Checks if the value of the field is ONE

pub fn is_two(&self) -> bool

Checks if the value of the field is TWO