Enum stm32f103xx::gpioa::crl::CNF0R

pub enum CNF0R {
    PUSH,
    OPEN,
    ALTPUSH,
    ALTOPEN,
}

Possible values of the field CNF0

Variants
PUSH

Push-Pull mode

OPEN

Open Drain-Mode

ALTPUSH

Alternate Function Push-Pull Mode

ALTOPEN

Alternate Function Open-Drain Mode

Methods

impl CNF0R

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_push(&self) -> bool

Checks if the value of the field is PUSH

pub fn is_open(&self) -> bool

Checks if the value of the field is OPEN

pub fn is_alt_push(&self) -> bool

Checks if the value of the field is ALTPUSH

pub fn is_alt_open(&self) -> bool

Checks if the value of the field is ALTOPEN