Module stm32f103xx::rcc::bdcr

Backup domain control register (RCC_BDCR)

Structs

BDRSTR
Value of the field

LSEBYPR
Value of the field

LSEONR
Value of the field

LSERDYR
Value of the field

R
Value read from the register

W
Value to write to the register

_BDRSTW
Proxy

_LSEBYPW
Proxy

_LSEONW
Proxy

_RTCENW
Proxy

_RTCSELW
Proxy

Enums

RTCSELR
Possible values of the field RTCSEL

RTCSELW
Values that can be written to the field RTCSEL

Type Definitions

RTCENR
Possible values of the field RTCEN

RTCENW
Values that can be written to the field RTCEN