Module stm32f103xx::rcc::cfgr

Clock configuration register (RCC_CFGR)

Structs

MCOR
Value of the field

R
Value read from the register

W
Value to write to the register

_ADCPREW
Proxy

_HPREW
Proxy

_MCOW
Proxy

_PLLMULW
Proxy

_PLLSRCW
Proxy

_PLLXTPREW
Proxy

_PPRE1W
Proxy

_PPRE2W
Proxy

_SWW
Proxy

_USBPREW
Proxy

Enums

ADCPRER
Possible values of the field ADCPRE

ADCPREW
Values that can be written to the field ADCPRE

HPRER
Possible values of the field HPRE

HPREW
Values that can be written to the field HPRE

PLLMULR
Possible values of the field PLLMUL

PLLMULW
Values that can be written to the field PLLMUL

PLLSRCR
Possible values of the field PLLSRC

PLLSRCW
Values that can be written to the field PLLSRC

PLLXTPRER
Possible values of the field PLLXTPRE

PLLXTPREW
Values that can be written to the field PLLXTPRE

PPRE1R
Possible values of the field PPRE1

PPRE1W
Values that can be written to the field PPRE1

PPRE2R
Possible values of the field PPRE2

PPRE2W
Values that can be written to the field PPRE2

SWR
Possible values of the field SW

SWSR
Possible values of the field SWS

SWW
Values that can be written to the field SW

USBPRER
Possible values of the field USBPRE

USBPREW
Values that can be written to the field USBPRE