Enum stm32f103xx::rcc::cfgr::HPRER

pub enum HPRER {
    NODIV,
    DIV2,
    DIV4,
    DIV8,
    DIV16,
    DIV64,
    DIV128,
    DIV256,
    DIV512,
    _Reserved(u8),
}

Possible values of the field HPRE

Variants
NODIV

SYSCLK not divided

DIV2

SYSCLK divided by 2

DIV4

SYSCLK divided by 4

DIV8

SYSCLK divided by 8

DIV16

SYSCLK divided by 16

DIV64

SYSCLK divided by 64

DIV128

SYSCLK divided by 128

DIV256

SYSCLK divided by 256

DIV512

SYSCLK divided by 512

_Reserved(u8)

Reserved

Methods

impl HPRER

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_no_div(&self) -> bool

Checks if the value of the field is NODIV

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256

pub fn is_div512(&self) -> bool

Checks if the value of the field is DIV512