Enum stm32f103xx::rcc::cfgr::PLLXTPRER

pub enum PLLXTPRER {
    NODIV,
    DIV2,
}

Possible values of the field PLLXTPRE

Variants
NODIV

HSE not divided

DIV2

HSE divided by 2

Methods

impl PLLXTPRER

pub fn bit_is_clear(&self) -> bool

Returns true if the bit is clear (0)

pub fn bit_is_set(&self) -> bool

Returns true if the bit is set (1)

pub fn bit(&self) -> bool

Value of the field as raw bits

pub fn is_no_div(&self) -> bool

Checks if the value of the field is NODIV

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2