Enum stm32f103xx::rcc::cfgr::PPRE1R

pub enum PPRE1R {
    NODIV,
    DIV2,
    DIV4,
    DIV8,
    DIV16,
    _Reserved(u8),
}

Possible values of the field PPRE1

Variants
NODIV

HCLK not divided

DIV2

HCLK divided by 2

DIV4

HCLK divided by 4

DIV8

HCLK divided by 8

DIV16

HCLK divided by 16

_Reserved(u8)

Reserved

Methods

impl PPRE1R

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_no_div(&self) -> bool

Checks if the value of the field is NODIV

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16