Enum stm32f103xx::rcc::cr::PLLONR

pub enum PLLONR {
    DISABLED,
    ENABLED,
}

Possible values of the field PLLON

Variants
DISABLED

PLL disabled

ENABLED

Pll enabled

Methods

impl PLLONR

pub fn bit_is_clear(&self) -> bool

Returns true if the bit is clear (0)

pub fn bit_is_set(&self) -> bool

Returns true if the bit is set (1)

pub fn bit(&self) -> bool

Value of the field as raw bits

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_enabled(&self) -> bool

Checks if the value of the field is ENABLED