Module stm32f103xx::spi1::cr1

control register 1

Structs

BIDIMODER
Value of the field

BIDIOER
Value of the field

CPHAR
Value of the field

CPOLR
Value of the field

CRCENR
Value of the field

CRCNEXTR
Value of the field

LSBFIRSTR
Value of the field

R
Value read from the register

RXONLYR
Value of the field

SSIR
Value of the field

SSMR
Value of the field

W
Value to write to the register

_BIDIMODEW
Proxy

_BIDIOEW
Proxy

_BRW
Proxy

_CPHAW
Proxy

_CPOLW
Proxy

_CRCENW
Proxy

_CRCNEXTW
Proxy

_DFFW
Proxy

_LSBFIRSTW
Proxy

_MSTRW
Proxy

_RXONLYW
Proxy

_SPEW
Proxy

_SSIW
Proxy

_SSMW
Proxy

Enums

BRR
Possible values of the field BR

BRW
Values that can be written to the field BR

DFFR
Possible values of the field DFF

DFFW
Values that can be written to the field DFF

MSTRR
Possible values of the field MSTR

MSTRW
Values that can be written to the field MSTR

Type Definitions

SPER
Possible values of the field SPE

SPEW
Values that can be written to the field SPE