Enum stm32f103xx::spi1::cr1::BRR

pub enum BRR {
    DIV2,
    DIV4,
    DIV8,
    DIV16,
    DIV32,
    DIV64,
    DIV128,
    DIV256,
}

Possible values of the field BR

Variants
DIV2

f_PCLK/2

DIV4

f_PCLK/4

DIV8

f_PCLK/8

DIV16

f_PCLK/16

DIV32

f_PCLK/32

DIV64

f_PCLK/64

DIV128

f_PCLK/128

DIV256

f_PCLK/256

Methods

impl BRR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256