Enum stm32f103xx::tim1::ccmr1_output::OC1MR

pub enum OC1MR {
    FROZEN,
    SETACTIVE,
    SETINACTIVE,
    TOGGLE,
    FORCEINACTIVE,
    FORCEACTIVE,
    PWM1,
    PWM2,
}

Possible values of the field OC1M

Variants
FROZEN

The comparison between the output compare register TIMx_CCRy and the counter TIMx_CNT has no effect on the outputs(

SETACTIVE

Set channel y to active level on match. OCyREF signal is forced high when the counter TIMx_CNT matches the capture/compare register y (TIMx_CCRy).

SETINACTIVE

Set channel y to inactive level on match. OCyREF signal is forced low when the counter TIMx_CNT matches the capture/compare register y (TIMx_CCRy).

TOGGLE

OCyREF toggles when TIMx_CNT=TIMx_CCRy.

FORCEINACTIVE

OCyREF is forced low.

FORCEACTIVE

OCyREF is forced high.

PWM1

In upcounting, channel 1 is active as long as TIMx_CNTTIMx_CCRy else active (OCyREF=1).

PWM2

In upcounting, channel y is inactive as long as TIMx_CNTTIMx_CCRy else inactive.

Methods

impl OC1MR

fn bits(&self) -> u8

Value of the field as raw bits

fn is_frozen(&self) -> bool

Checks if the value of the field is FROZEN

fn is_set_active(&self) -> bool

Checks if the value of the field is SETACTIVE

fn is_set_inactive(&self) -> bool

Checks if the value of the field is SETINACTIVE

fn is_toggle(&self) -> bool

Checks if the value of the field is TOGGLE

fn is_force_inactive(&self) -> bool

Checks if the value of the field is FORCEINACTIVE

fn is_force_active(&self) -> bool

Checks if the value of the field is FORCEACTIVE

fn is_pwm1(&self) -> bool

Checks if the value of the field is PWM1

fn is_pwm2(&self) -> bool

Checks if the value of the field is PWM2