Struct stm32f103xx::tim1::ccmr1_output::_OC1MW

pub struct _OC1MW<'a> { /* fields omitted */ }

Proxy

Methods

impl<'a> _OC1MW<'a>

fn variant(self, variant: OC1MW) -> &'a mut W

Writes variant to the field

fn frozen(self) -> &'a mut W

The comparison between the output compare register TIMx_CCRy and the counter TIMx_CNT has no effect on the outputs(

fn set_active(self) -> &'a mut W

Set channel y to active level on match. OCyREF signal is forced high when the counter TIMx_CNT matches the capture/compare register y (TIMx_CCRy).

fn set_inactive(self) -> &'a mut W

Set channel y to inactive level on match. OCyREF signal is forced low when the counter TIMx_CNT matches the capture/compare register y (TIMx_CCRy).

fn toggle(self) -> &'a mut W

OCyREF toggles when TIMx_CNT=TIMx_CCRy.

fn force_inactive(self) -> &'a mut W

OCyREF is forced low.

fn force_active(self) -> &'a mut W

OCyREF is forced high.

fn pwm1(self) -> &'a mut W

In upcounting, channel 1 is active as long as TIMx_CNTTIMx_CCRy else active (OCyREF=1).

fn pwm2(self) -> &'a mut W

In upcounting, channel y is inactive as long as TIMx_CNTTIMx_CCRy else inactive.

fn bits(self, value: u8) -> &'a mut W

Writes raw bits to the field