Enum stm32f103xx::tim1::cr1::DIRR

pub enum DIRR {
    UP,
    DOWN,
}

Possible values of the field DIR

Variants
UP

Up

DOWN

Down

Methods

impl DIRR

pub fn bit_is_clear(&self) -> bool

Returns true if the bit is clear (0)

pub fn bit_is_set(&self) -> bool

Returns true if the bit is set (1)

pub fn bit(&self) -> bool

Value of the field as raw bits

pub fn is_up(&self) -> bool

Checks if the value of the field is UP

pub fn is_down(&self) -> bool

Checks if the value of the field is DOWN