Enum stm32f103xx::tim1::smcr::SMSR

pub enum SMSR {
    DISABLED,
    ENCODERTI2,
    ENCODERTI1,
    ENCODERTI1TI2,
    RESET,
    GATED,
    TRIGGER,
    EXTERNAL,
}

Possible values of the field SMS

Variants
DISABLED

Counter disabled

ENCODERTI2

Encoder mode, count up/down on TI2FP1

ENCODERTI1

Encoder mode, count up/down on TI1FP2

ENCODERTI1TI2

Encoder mode, count up/down on both TI1FP1 and TI2FP2

RESET

Rising edge of the selected trigger input (TRGI) reinitializes the counter

GATED

The counter clock is enabled when the trigger input (TRGI) is high

TRIGGER

The counter starts at a rising edge of the trigger TRGI

EXTERNAL

Rising edges of the selected trigger (TRGI) clock the counter

Methods

impl SMSR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_disabled(&self) -> bool

Checks if the value of the field is DISABLED

pub fn is_encoder_ti2(&self) -> bool

Checks if the value of the field is ENCODERTI2

pub fn is_encoder_ti1(&self) -> bool

Checks if the value of the field is ENCODERTI1

pub fn is_encoder_ti1ti2(&self) -> bool

Checks if the value of the field is ENCODERTI1TI2

pub fn is_reset(&self) -> bool

Checks if the value of the field is RESET

pub fn is_gated(&self) -> bool

Checks if the value of the field is GATED

pub fn is_trigger(&self) -> bool

Checks if the value of the field is TRIGGER

pub fn is_external(&self) -> bool

Checks if the value of the field is EXTERNAL