Enum stm32f103xx::tim1::smcr::TSR

pub enum TSR {
    ITR0,
    ITR1,
    ITR2,
    ITR3,
    TI1F_ED,
    TI1FP1,
    TI2FP2,
    ETRF,
}

Possible values of the field TS

Variants
ITR0

Internal Trigger 0 (ITR0)

ITR1

Internal Trigger 1 (ITR1)

ITR2

Internal Trigger 2 (ITR2)

ITR3

Internal Trigger 3 (ITR3)

TI1F_ED

TI1 Edge Detector

TI1FP1

Filtered Timer Input 1

TI2FP2

Filtered Timer Input 2

ETRF

External Trigger input

Methods

impl TSR

fn bits(&self) -> u8

Value of the field as raw bits

fn is_itr0(&self) -> bool

Checks if the value of the field is ITR0

fn is_itr1(&self) -> bool

Checks if the value of the field is ITR1

fn is_itr2(&self) -> bool

Checks if the value of the field is ITR2

fn is_itr3(&self) -> bool

Checks if the value of the field is ITR3

fn is_ti1f_ed(&self) -> bool

Checks if the value of the field is TI1F_ED

fn is_ti1fp1(&self) -> bool

Checks if the value of the field is TI1FP1

fn is_ti2fp2(&self) -> bool

Checks if the value of the field is TI2FP2

fn is_etrf(&self) -> bool

Checks if the value of the field is ETRF