Enum stm32f103xx::tim1::sr::UIFR

pub enum UIFR {
    NOUPDATE,
    PENDING,
}

Possible values of the field UIF

Variants
NOUPDATE

No update occurred

PENDING

Update interrupt pending

Methods

impl UIFR

pub fn bit_is_clear(&self) -> bool

Returns true if the bit is clear (0)

pub fn bit_is_set(&self) -> bool

Returns true if the bit is set (1)

pub fn bit(&self) -> bool

Value of the field as raw bits

pub fn is_no_update(&self) -> bool

Checks if the value of the field is NOUPDATE

pub fn is_pending(&self) -> bool

Checks if the value of the field is PENDING