Struct stm32f103xx::tim2::smcr::_SMSW

pub struct _SMSW<'a> { /* fields omitted */ }

Proxy

Methods

impl<'a> _SMSW<'a>

pub fn variant(self, variant: SMSW) -> &'a mut W

Writes variant to the field

pub fn disabled(self) -> &'a mut W

Counter disabled

pub fn encoder_ti2(self) -> &'a mut W

Encoder mode, count up/down on TI2FP1

pub fn encoder_ti1(self) -> &'a mut W

Encoder mode, count up/down on TI1FP2

pub fn encoder_ti1ti2(self) -> &'a mut W

Encoder mode, count up/down on both TI1FP1 and TI2FP2

pub fn reset(self) -> &'a mut W

Rising edge of the selected trigger input (TRGI) reinitializes the counter

pub fn gated(self) -> &'a mut W

The counter clock is enabled when the trigger input (TRGI) is high

pub fn trigger(self) -> &'a mut W

The counter starts at a rising edge of the trigger TRGI

pub fn external(self) -> &'a mut W

Rising edges of the selected trigger (TRGI) clock the counter

pub fn bits(self, value: u8) -> &'a mut W

Writes raw bits to the field