Enum stm32f103xx::rcc::cfgr::PLLMULR

pub enum PLLMULR {
    MUL2,
    MUL3,
    MUL4,
    MUL5,
    MUL6,
    MUL7,
    MUL8,
    MUL9,
    MUL10,
    MUL11,
    MUL12,
    MUL13,
    MUL14,
    MUL15,
    MUL16,
    _Reserved(u8),
}

Possible values of the field PLLMUL

Variants
MUL2

PLL input clock x 2

MUL3

PLL input clock x 3

MUL4

PLL input clock x 4

MUL5

PLL input clock x 5

MUL6

PLL input clock x 6

MUL7

PLL input clock x 7

MUL8

PLL input clock x 8

MUL9

PLL input clock x 9

MUL10

PLL input clock x 10

MUL11

PLL input clock x 11

MUL12

PLL input clock x 12

MUL13

PLL input clock x 13

MUL14

PLL input clock x 14

MUL15

PLL input clock x 15

MUL16

PLL input clock x 16

_Reserved(u8)

Reserved

Methods

impl PLLMULR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_mul2(&self) -> bool

Checks if the value of the field is MUL2

pub fn is_mul3(&self) -> bool

Checks if the value of the field is MUL3

pub fn is_mul4(&self) -> bool

Checks if the value of the field is MUL4

pub fn is_mul5(&self) -> bool

Checks if the value of the field is MUL5

pub fn is_mul6(&self) -> bool

Checks if the value of the field is MUL6

pub fn is_mul7(&self) -> bool

Checks if the value of the field is MUL7

pub fn is_mul8(&self) -> bool

Checks if the value of the field is MUL8

pub fn is_mul9(&self) -> bool

Checks if the value of the field is MUL9

pub fn is_mul10(&self) -> bool

Checks if the value of the field is MUL10

pub fn is_mul11(&self) -> bool

Checks if the value of the field is MUL11

pub fn is_mul12(&self) -> bool

Checks if the value of the field is MUL12

pub fn is_mul13(&self) -> bool

Checks if the value of the field is MUL13

pub fn is_mul14(&self) -> bool

Checks if the value of the field is MUL14

pub fn is_mul15(&self) -> bool

Checks if the value of the field is MUL15

pub fn is_mul16(&self) -> bool

Checks if the value of the field is MUL16