Module stm32f103xx::sdio

Secure digital input/output interface

Modules

arg
Bits 31:0 = : Command argument

clkcr
SDI clock control register (SDIO_CLKCR)

cmd
SDIO command register (SDIO_CMD)

dcount
Bits 24:0 = DATACOUNT: Data count value

dctrl
SDIO data control register (SDIO_DCTRL)

dlen
Bits 24:0 = DATALENGTH: Data length value

dtimer
Bits 31:0 = DATATIME: Data timeout period

fifo
bits 31:0 = FIFOData: Receive and transmit FIFO data

fifocnt
Bits 23:0 = FIFOCOUNT: Remaining number of words to be written to or read from the FIFO

icr
SDIO interrupt clear register (SDIO_ICR)

mask
SDIO mask register (SDIO_MASK)

power
Bits 1:0 = PWRCTRL: Power supply control bits

resp2
Bits 31:0 = CARDSTATUS2

resp3
Bits 31:0 = CARDSTATUS3

resp4
Bits 31:0 = CARDSTATUS4

respcmd
SDIO command register

respi1
Bits 31:0 = CARDSTATUS1

sta
SDIO status register (SDIO_STA)

Structs

ARG
Bits 31:0 = : Command argument

CLKCR
SDI clock control register (SDIO_CLKCR)

CMD
SDIO command register (SDIO_CMD)

DCOUNT
Bits 24:0 = DATACOUNT: Data count value

DCTRL
SDIO data control register (SDIO_DCTRL)

DLEN
Bits 24:0 = DATALENGTH: Data length value

DTIMER
Bits 31:0 = DATATIME: Data timeout period

FIFO
bits 31:0 = FIFOData: Receive and transmit FIFO data

FIFOCNT
Bits 23:0 = FIFOCOUNT: Remaining number of words to be written to or read from the FIFO

ICR
SDIO interrupt clear register (SDIO_ICR)

MASK
SDIO mask register (SDIO_MASK)

POWER
Bits 1:0 = PWRCTRL: Power supply control bits

RESP2
Bits 31:0 = CARDSTATUS2

RESP3
Bits 31:0 = CARDSTATUS3

RESP4
Bits 31:0 = CARDSTATUS4

RESPCMD
SDIO command register

RESPI1
Bits 31:0 = CARDSTATUS1

RegisterBlock
Register block

STA
SDIO status register (SDIO_STA)