Struct stm32f103xx::SCB

pub struct SCB { /* fields omitted */ }

System Control Block

Methods

impl SCB

pub fn enable_icache(&mut self)

Enables I-Cache if currently disabled

pub fn disable_icache(&mut self)

Disables I-Cache if currently enabled

pub fn icache_enabled() -> bool

Returns whether the I-Cache is currently enabled

pub fn invalidate_icache(&mut self)

Invalidates I-Cache

pub fn enable_dcache(&mut self, cpuid: &mut CPUID)

Enables D-cache if currently disabled

pub fn disable_dcache(&mut self, cpuid: &mut CPUID)

Disables D-cache if currently enabled

pub fn dcache_enabled() -> bool

Returns whether the D-Cache is currently enabled

pub fn clean_dcache(&mut self, cpuid: &mut CPUID)

Cleans D-cache

pub fn clean_invalidate_dcache(&mut self, cpuid: &mut CPUID)

Cleans and invalidates D-cache

pub fn invalidate_dcache_by_address(&mut self, addr: usize, size: usize)

Invalidates D-cache by address

addr: the address to invalidate size: size of the memory block, in number of bytes

Invalidates cache starting from the lowest 32-byte aligned address represented by addr, in blocks of 32 bytes until at least size bytes have been invalidated.

pub fn clean_dcache_by_address(&mut self, addr: usize, size: usize)

Cleans D-cache by address

addr: the address to clean size: size of the memory block, in number of bytes

Cleans cache starting from the lowest 32-byte aligned address represented by addr, in blocks of 32 bytes until at least size bytes have been cleaned.

pub fn clean_invalidate_dcache_by_address(&mut self, addr: usize, size: usize)

Cleans and invalidates D-cache by address

addr: the address to clean and invalidate size: size of the memory block, in number of bytes

Cleans and invalidates cache starting from the lowest 32-byte aligned address represented by addr, in blocks of 32 bytes until at least size bytes have been cleaned and invalidated.

impl SCB

pub fn ptr() -> *const RegisterBlock

Returns a pointer to the register block