Enum stm32f103xx::tim1::ccmr1_output::OC1MW

pub enum OC1MW {
    FROZEN,
    SETACTIVE,
    SETINACTIVE,
    TOGGLE,
    FORCEINACTIVE,
    FORCEACTIVE,
    PWM1,
    PWM2,
}

Values that can be written to the field OC1M

Variants
FROZEN

The comparison between the output compare register TIMx_CCRy and the counter TIMx_CNT has no effect on the outputs(

SETACTIVE

Set channel y to active level on match. OCyREF signal is forced high when the counter TIMx_CNT matches the capture/compare register y (TIMx_CCRy).

SETINACTIVE

Set channel y to inactive level on match. OCyREF signal is forced low when the counter TIMx_CNT matches the capture/compare register y (TIMx_CCRy).

TOGGLE

OCyREF toggles when TIMx_CNT=TIMx_CCRy.

FORCEINACTIVE

OCyREF is forced low.

FORCEACTIVE

OCyREF is forced high.

PWM1

In upcounting, channel 1 is active as long as TIMx_CNT<TIMx_CCRy else inactive. In downcounting, channel 1 is inactive (OCyREF=‘0) as long as TIMx_CNT>TIMx_CCRy else active (OCyREF=1).

PWM2

In upcounting, channel y is inactive as long as TIMx_CNT<TIMx_CCRy else active. In downcounting, channel y is active as long as TIMx_CNT>TIMx_CCRy else inactive.